Add files via upload
This is the infrastructure design we have created on the cloud path. Each service stands alone and uses resources according to its needs. The mobile app communicates with the server on the google compute engine via a REST API using ExpressJS. Service for authentication using Node JS implemented in Compute Engine. The data from the server is stored in the MongoDB database and google cloud storage. Then Model is implemented in google Compute Engine with tensorflowjs.